// apps/web/generated/api/models/dashboardStatisticsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStatisticsResponse {
  total_users: number;
  active_projects: number;
  total_freelancers: number;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  total_revenue: string;
}